requirements.txt
Flask
pymongo
python-dotenv
gunicorn
Flask-PyMongo
python-slugify
